+++ src/main/java/com/example/HMSPHARMACY/MODEL/Sales.java
@@ -10,7 +10,7 @@ public class Sales {
     Long id;
 
     String productName;
-    Double productQuantity ;
+    Long productQuantity ;
     Double productPrice;
     String status;
 
@@ -50,11 +50,11 @@ public class Sales {
         this.productName = productName;
     }
 
-    public Double getProductQuantity() {
+    public Long getProductQuantity() {
         return productQuantity;
     }
 
-    public void setProductQuantity(Double productQuantity) {
+    public void setProductQuantity(Long productQuantity) {
         this.productQuantity = productQuantity;
     }
